preemptive numbers

+++ src/misc.hpp
@@ -4,22 +4,23 @@
 #include <string>
 #include <fstream>
 #include <stdexcept>
+#include <cstdint>
 
 class Misc {
 private:
-    std::string input;
-    size_t pos;
-    size_t line;
-    size_t column;
+	std::string input;
+    	uint64_t pos;
+    	uint64_t line;
+	uint64_t column;
 
 public:
-    Misc(const std::string& filename);
-    Misc(const std::string& src, size_t start_line = 1, size_t start_column = 1);
-    char peek() const;
-    char get();
-    bool eof() const;
-    size_t getLine() const { return line; } 
-    size_t getColumn() const { return column; }
-    std::string getRemaining() const;
+	Misc(const std::string& filename);
+	Misc(const std::string& src, uint64_t start_line = 1, uint64_t start_column = 1);
+	char peek() const;
+    	char get();
+    	bool eof() const;
+    	uint64_t getLine() const { return line; } 
+    	uint64_t getColumn() const { return column; }
+    	std::string getRemaining() const;
 };
 #endif
